feat(arch): drafting bot settling flow
summary: the order book is comrpsied of effective queues of trades, identified via a mapping with their unique tradeId. once queued and ready to stream, a bot calls a settling function which loops to predefined thesholds wrt gas limits, and provides for the function of the protocol as trades roll in.