feat(arch): drafting the keeper to DEX monitor
summary: the frontend will fetch data directly from DEXs or an indexer (The Graph e.g.), process them and provide the processed datas as trade parameters for entry of trades, specifically for the estimation of gas fees to be included in the total fees taken by a bot.

Currently, the architecture relies on on-chain calls to actually calculate stream size dynamically. The algorithm is to be defined under documentation/smart-contracts/StreamDaemon.sol.md