feat(arch) drafting functional requirements for the Executor contract
summary: this is a first draft contract diagram derived from the flows in the adjacent directory. it describes the functional requirements of the contract and should be used as a reference point during specification and implementation. in essence, the executor settles trades against a given DEX, architected via function calls to the StreamDaemon.