feat(arch): finalising process of trade construction
summary: this is a full process flow for the STREAM functionality of the protocol from the perspective of a user placing a trade. It utilises the full front end, middleware and smart contract suite.

N.B. the diagram doesn't display flows trades that are executed in a single stream, or consider the taking of fees