feat(arch): completing functional arch flow for StreamDaemon
summary: the StreamDaemon contract has functional requirements that have been derived from the diagrams defined in /diagrams/flows. These are used as a basis to outline functional requirements of the StreamDaemon contrat, displayed in this diagram. Note that this is not a complete specification. Other requirements will be defined more explicitly in /documentation