feat(contracts): draft for StreamDaemon complete
summary: this contract is responsible for returning DEX's with high volume given a token pair and utilising that data feed to generate a sweet spot for the number of streams given any trade entry. This commit features a barebones contract, in line with the arch diagram, which is designed to achieve this. It is a rough, untested draft. The next step requires specific variables, getters and setters for the contract to be written. Post that, testing interactions through a fork or similar will provide for suitable integration tests.