feat(contracts): removing execution and state bearing logic from Executor
summary: the Executor previously performed logic calculations in the settling of a trade. This new version is totally stripped out, with state and relative logics now moved to the StreeamDaemon. This commit presents a relatively empty Executor, where execution dedicated logic will be implemented in the settling of trades via `delegatecall` from Core.