feat(contracts): adding trades to Core contract storage
summary: previously, Core did not feature any dstorage methods for trades - rather, the functionality of executing a trade of arbitrary volume was implemented to execute across a variety of (user defined) DEXs. This commit sees the first implementation of long-term storage for user submitted trades. The execution logic from a maintainer;'s perspective is yet to be integrated.