ferat(contracts): implementing the placing of trades from EOA
summary: when a user (eoa or smart wallet) places a trade, their trade is built in memory, and a single streram is executed. If trade volume remains after this execution (i.e. streamCount != 0), the trade is stored in contract in the effective queue/order book. If streamCount = 0, the funds are directly transferred out and nothing is stored.

N.B.
1. There are no events yet!
2. No fees are taken at this stage!