fix(contracts): fixing scaling errors on the sweetSpot algo
summary: the algo now works, with good values coming out and tests passing. This commit features updates to the trade entry and stream calculation flow, and should support in affect the placing of trades in the contract.

N.B. the execution against a DEX is yet to be fully tested