feat(contracts): addding Balancer and Sushiswap routing to _executeStream
summary: now, a bot is able to execute streams against a DEX as dictated by the StreamDaemon via calling executeTrades and passing a pairId. The sweet spot and appropriate DEX is found and now, with this commit, Core is able to interact with any of the Owner dictated DEXs in settling a stream from any trade.

Edge cases are also handled in sweetSpot evalution, ensuring trades are fully settled and closed.

A comprehensive suite of tests is yet to be completed for this feature(s).