feat(contracts): drafting instasettle
summary: instasettle allows a trader to settle any pending transactions at a BPS loss (for the trade owner) to provide for slipppage and front running free insta trade settling. This commit featured a draft version of this functionality(s). It's unit tests are to follow and must be considered draft.