feat(tests): completing the test suite for placeTrade()
summary: this commit features a functioning single paradigm test suite for the placement of a trade. This action takes a user's trade volume, stores metadata in the form of a trade, executes a single stream from that trade, caching critical data along the way, and finally returns the executed amounts to Core. It is a single full functional flow for the architecture. Comments are still included and redundant code has not ben removed.