feat(contracts): providing for execution across multiple DEXs
summary: our initial contracts were built to solely interact with UniswapV3. This commit features elements that allow the system to interact dynamically with any 3 of the main DEXs, being Uniswap v2, v3 and Sushiswap. Note that this is not a production model, since the abstraction of a DEX-agnostic approach is yet to come.