feat(tests): completing test suites for trade placement across DEXs
summary: the previous contracts commit featured an abstraction of DEX logics to a registry. This commit features an updated set of test suites, allowing for the placement of trades and the execution of a single stream. N.B. there are bugs are this should not be considered production ready. Invariants and process tests are incoming.