feat(contracts): implementing Instasettle assertions and accompanying test suite
summary: the previous architecture feature a drafted untested instasettle functionality, broken down into two functions in Core. This commit features a compiled singular function endpoint for the function call, as well as an accompanying lkimited test suite. The feature is not working in its entirey in this commit.

N.B. This commit also features an updated configuration to the package.json to accomodate the deployment and script testing of the protocol on a local fork via foundry / anvil.