feat(tests): adding local (anvil) testing framework
summary: the previous architecture ran tests against a remote RPC fork of mainnet. This meant that only a singular snapshot was being used to test the architecture. This commit features testing suites for remote (in this case, a buildbear fork - although this could be replaced with whichever a dev sees fit) and local forked environments via `npm run test`.