feat(contracts): deprecating botGasAllowance
summary: until now, trades contained metadata that was iteratively used to assess the gas usage on a trade. This was in place to protect bots from high gas spends and to reduce the impacet of dud transactions. This commit sees the removal of this aspect of trades, remitted on the basis that (a) the tradeAttempts counter is enough given any dud trade at this stage and (b) gas can be manipulated and may provide a vector for attack and (c) on the assumption that any bot's BPS allowance effectively provides for a buffer against tx execution, approximated at this stage to be ~40k gas per stream