feat(contracts): completing configurations for mainnet deployment
summary: the previous codebase featured logs ad lacked files and scripts to accomodate deployment to mainnet. This commit features a cleaned codebase, with adaptations made to accomodate Infura as our RPC provider and guides written in markdown to assist the addition of new DEXs by the team and/or community. The old code base, featuring logs for development purposes, exists on a branch mk1-final-draft-with-logs