feat(bot): creating an intuitive monitor for local and cron use
summary: the previous 'bot' was designed to explicitly run trades. It operated, however lacked intuitive logging and features allowing operators of the bot to monitor progress. This commit features a trimmed down, more intuitive version of the bot, which displays historical and current data, properly caching both and utilises the cached data to further execute trades and asynchronously update the localData.json (SC state) as runs are executed.